fix(sdk): harden scenario snapshots
